Line follows the cursor
So I got the cursor and line to move together. Issue I do not like how this is currently done
as I am iterating through the entire list every time, this is not optimal.

NOTE TO SELF:
I think I need to add the line to the list on enter.
I will mot likely need a column object to help determine exact location
will need to figure out how to push in the middle of the line vector